restrucuturing employee-facing file structuring for the HRMS
